Detect unicode characters in python3 friendly way.
Python 3 removed the decode function for str and encode for bytes and
all str are unicode. Some of the logic around detecting unicode and
validating fields works differently. For example python 3 detects that a
unicode character is not a valid character in the node_name validation.
When python 2 runs this same code, it doesn't catch it until the
encode('ascii') try on line 226. This means the encoding error will only
be shown for python 2, but from a high level the code is equivalent.